separating large-file-copy process into a static object with its own actor system and pure default (local) configuration
